% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratios.R
\name{couscous}
\alias{couscous}
\title{Calculates how much water you need for your couscous}
\usage{
couscous(quantity, unit = c("g", "cups"))
}
\arguments{
\item{quantity}{How much couscous in g or cups}

\item{unit}{Enter "g" for grams or "cups" for cups}
}
\value{
Amount of water you need
}
\description{
Calculates how much water you need for your couscous
}

