% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grains.R
\docType{package}
\name{grains}
\alias{grains}
\alias{grains-package}
\title{grains: How much water do you need for your grains?}
\description{
\code{grains} contains the following functions:

\link{couscous}
\link{quinoa}
\link{rice}
}
\details{
Calculates how much water you need for your grains. I am working on the grain repetoir and ratios
and accept no responsibility for failed cooking attempts.
}
\author{
Emily Humble (emily.lhumble@gmail.com)
}

