% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratios.R
\name{quinoa}
\alias{quinoa}
\title{Calculates how much water you need for your quinoa}
\usage{
quinoa(quantity, unit = c("g", "cups"))
}
\arguments{
\item{quantity}{How much quinoa in g or cups}

\item{unit}{Enter "g" for grams or "cups" for cups}
}
\value{
Amount of water you need
}
\description{
Calculates how much water you need for your quinoa
}

