% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratios.R
\name{rice}
\alias{rice}
\title{Calculates how much water you need for your rice}
\usage{
rice(quantity, unit = c("g", "cups"))
}
\arguments{
\item{quantity}{How much rice in g or cups}

\item{unit}{Enter "g" for grams or "cups" for cups}
}
\value{
Amount of water you need
}
\description{
Calculates how much water you need for your rice
}

